ml-models/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.2.0+cu118
torchvision==0.17.0+cu118
transformers==4.35.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.4
python-multipart==0.0.6
redis==5.0.1
openai==1.40.0
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
